Changed All games: cvar hud-status-size is now defined as floating point.

+++ doomsday/plugins/jdoom/include/d_config.h
@@ -125,7 +125,7 @@ typedef struct jdoom_config_s {
     byte            coopRespawnItems;
     byte            respawnMonstersNightmare;
 
-    int             statusbarScale;
+    float           statusbarScale;
     float           statusbarOpacity;
     float           statusbarCounterAlpha;
